Setup Postgres migrations with sqlx
Sqlx seems to be a good choice for an async database driver (recommended
in zero2prod), so let's give it a go.

The fact that it 1) manages migrations and 2) lets us do that in plain
SQL is a win in my books; I'm tired of learning a DSL for migrations.